# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Covalence API",
        "description": "Hybrid GraphRAG knowledge engine API",
        "license": {
            "name": "MIT",
            "identifier": "MIT"
        },
        "version": "0.1.0"
    },
    "paths": {
...
        "/nodes/merge": {
            "post": {
                "tags": [
                    "nodes"
                ],
                "summary": "Merge multiple nodes into a target.",
                "operationId": "merge_nodes",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MergeNodesRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Merge completed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MergeNodesResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Not yet implemented"
                    }
                }
            }
        },
        "/nodes/resolve": {
            "post": {
                "tags": [
                    "nodes"
                ],
                "summary": "Resolve a name to a node.",
                "operationId": "resolve_node",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ResolveNodeRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Resolved node",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NodeResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No matching node"
                    }
                }
            }
        },
        "/nodes/{id}": {
            "get": {
                "tags": [
                    "nodes"
                ],
                "summary": "Get a node by ID.",
                "operationId": "get_node",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Node ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Node found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NodeResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Node not found"
                    }
                }
            }
        },
        "/nodes/{id}/neighborhood": {
            "get": {
                "tags": [
                    "nodes"
                ],
                "summary": "Get the neighborhood of a node.",
                "operationId": "get_neighborhood",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Node ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "hops",
                        "in": "query",
                        "description": "Number of hops (default 1).",
                        "required": false,
                        "schema": {
                            "type": [
                                "integer",
                                "null"
                            ],
                            "minimum": 0
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Neighbor nodes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/NodeResponse"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/nodes/{id}/provenance": {
            "get": {
                "tags": [
                    "nodes"
                ],
                "summary": "Get the provenance chain for a node.",
                "operationId": "get_provenance",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Node ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Provenance chain",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProvenanceResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/nodes/{id}/split": {
            "post": {
                "tags": [
                    "nodes"
                ],
                "summary": "Split a node.",
                "operationId": "split_node",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Node ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SplitNodeRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Split completed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SplitNodeResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Node not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "MergeNodesRequest": {
                "type": "object",
                "description": "Request body for merging nodes.",
                "required": [
                    "source_ids",
                    "target_id"
                ],
                "properties": {
                    "source_ids": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "description": "IDs of source nodes to merge from."
                    },
                    "target_id": {
                        "type": "string",
                        "format": "uuid",
                        "description": "ID of the target node to merge into."
                    }
                }
            },
            "MergeNodesResponse": {
                "type": "object",
                "description": "Response after merge.",
                "required": [
                    "audit_log_id"
                ],
                "properties": {
                    "audit_log_id": {
                        "type": "string",
                        "format": "uuid",
                        "description": "Audit log entry ID for the merge operation."
                    }
                }
            },
...
            "NodeResponse": {
                "type": "object",
                "description": "Response for a node entity.",
                "required": [
                    "id",
                    "canonical_name",
                    "node_type",
                    "properties",
                    "clearance_level",
                    "first_seen",
                    "last_seen",
                    "mention_count"
                ],
                "properties": {
                    "canonical_name": {
                        "type": "string"
                    },
                    "clearance_level": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "description": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "first_seen": {
                        "type": "string"
                    },
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "last_seen": {
                        "type": "string"
                    },
                    "mention_count": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "node_type": {
                        "type": "string"
                    },
                    "properties": {}
                }
            },
...
            "ProvenanceResponse": {
                "type": "object",
                "description": "Provenance chain response.",
                "required": [
                    "node_id",
                    "extraction_count",
                    "chunk_count",
                    "source_count"
                ],
                "properties": {
                    "chunk_count": {
                        "type": "integer",
                        "description": "Number of source chunks.",
                        "minimum": 0
                    },
                    "extraction_count": {
                        "type": "integer",
                        "description": "Number of extraction records.",
                        "minimum": 0
                    },
                    "node_id": {
                        "type": "string",
                        "format": "uuid",
                        "description": "Node ID."
                    },
                    "source_count": {
                        "type": "integer",
                        "description": "Number of originating sources.",
                        "minimum": 0
                    }
                }
            },
...
            "ResolveNodeRequest": {
                "type": "object",
                "description": "Request body for node resolution.",
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "description": "Name to resolve."
                    }
                }
            },
...
            "SplitNodeRequest": {
                "type": "object",
                "description": "Request for node split.",
                "required": [
                    "specs"
                ],
                "properties": {
                    "specs": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/SplitSpecRequest"
                        },
                        "description": "Specifications for each new node."
                    }
                }
            },
            "SplitNodeResponse": {
                "type": "object",
                "description": "Response after split.",
                "required": [
                    "node_ids"
                ],
                "properties": {
                    "node_ids": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "description": "IDs of the newly created nodes."
                    }
                }
            },
            "SplitSpecRequest": {
                "type": "object",
                "description": "A single split target specification.",
                "required": [
                    "name",
                    "node_type",
                    "edge_ids"
                ],
                "properties": {
                    "description": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "description": "Optional description."
                    },
                    "edge_ids": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "description": "Edge IDs to reassign to this new node."
                    },
                    "name": {
                        "type": "string",
                        "description": "Name for the new node."
                    },
                    "node_type": {
                        "type": "string",
                        "description": "Type for the new node."
                    }
                }
            }
        }
    },
...
}